// src/app/model/docs_societa_model.ts
export class Docs_societa_model {

  questionario: File = null;
  visuraCamerale: File = null;
  iso9001: File = null;
  polizzaRctRco: File = null;
  durc: File = null;
  bilancio: File = null;
  constructor(values: Object = {}) {
    //Constructor initialization
    Object.assign(this, values);
  }
}